feat(tabs): rename tabs component and extends props
