+++ src/atoms/index.ts
@@ -9,7 +9,7 @@ import Grid from './grid';
 import Icon from './icon';
 import Link from './link';
 import Password from './password';
-import TabPanel from './tab-panel';
+import Tabs from './tabs';
 import TextField from './text-field';
 import Typography from './typography';
 
@@ -25,7 +25,7 @@ export {
   Icon,
   Link,
   Password,
-  TabPanel,
+  Tabs,
   TextField,
   Typography,
 };
