+++ src/atoms/tabs.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
+import {Tabs as MuiTabs, Tab as MuiTab, TabsProps as MuiTabsProps } from '@mui/material';
 import Box from '@mui/material/Box';
 import CircularProgress from '@mui/material/CircularProgress';
-import Tab from '@mui/material/Tab';
-import Tabs from '@mui/material/Tabs';
-import { IconType, iconLookup } from '../atoms/icon-store';
+
+import { IconType, iconLookup } from './icon-store';
 
 export interface TabConfig {
   callBack?: () => void;
@@ -18,14 +18,14 @@ export interface TabConfig {
   value: number;
 }
 
-interface TabPanelProps {
+interface TabsProps extends MuiTabsProps {
   initialTabIndex: number;
   orientation?: 'horizontal' | 'vertical';
   scrollOnTabLength: number;
   tabs: TabConfig[];
 }
 
-function TabPanel({ tabs, orientation = 'horizontal', scrollOnTabLength, initialTabIndex }: TabPanelProps) {
+function Tabs({ tabs, orientation = 'horizontal', scrollOnTabLength, initialTabIndex }: TabsProps) {
   const initialTabValue = tabs.find(tab => tab.value === initialTabIndex)?.value || 0;
   const [value, setValue] = React.useState<number>(initialTabValue);
 
@@ -39,7 +39,7 @@ function TabPanel({ tabs, orientation = 'horizontal', scrollOnTabLength, initial
 
   return (
     <Box sx={{ display: 'flex', flexDirection: orientation === 'vertical' ? 'row' : 'column', width: '400px' }}>
-      <Tabs
+      <MuiTabs
         onChange={handleChange}
         orientation={orientation}
         scrollButtons="auto"
@@ -50,7 +50,7 @@ function TabPanel({ tabs, orientation = 'horizontal', scrollOnTabLength, initial
         {tabs.map((tab) => {
           const Icon = tab.icon ? iconLookup[tab.icon] : undefined;
           return (
-            <Tab
+            <MuiTab
               key={tab.value}
               disabled={tab.disabled}
               icon={Icon && <Icon />}
@@ -61,7 +61,7 @@ function TabPanel({ tabs, orientation = 'horizontal', scrollOnTabLength, initial
             />
           );
         })}
-      </Tabs>
+      </MuiTabs>
       <Box sx={{ flexGrow: 1, padding: 2 }}>
         <React.Suspense fallback={<CircularProgress />}>
           {tabs.find((tab) => tab.value === value)?.component}
@@ -71,4 +71,4 @@ function TabPanel({ tabs, orientation = 'horizontal', scrollOnTabLength, initial
   );
 }
 
-export default TabPanel;
+export default Tabs;
